remove trivial module causing import issues
